lint and file organization
