+++ scripts/start.js
@@ -30,8 +30,6 @@ const config = require('../config/webpack.config.dev');
 const createDevServerConfig = require('../config/webpackDevServer.config');
 const mockServer = require('./mock');
 
-mockServer();
-
 const useYarn = fs.existsSync(paths.yarnLockFile);
 const isInteractive = process.stdout.isTTY;
